move glm utils
